zkvm/merkle: factor out merkle proofs into separate module (#200)
Fixes #197
