+++ zkvm/src/lib.rs
@@ -8,6 +8,7 @@ mod constraints;
 mod contract;
 mod encoding;
 mod errors;
+mod merkle;
 mod ops;
 mod point_ops;
 mod predicate;
@@ -23,6 +24,7 @@ mod vm;
 pub use self::constraints::{Commitment, Constraint, Expression, Variable};
 pub use self::contract::{Contract, Input, PortableItem};
 pub use self::errors::VMError;
+pub use self::merkle::{MerkleItem, MerkleNeighbor, MerkleTree};
 pub use self::ops::{Instruction, Opcode, Program};
 pub use self::predicate::Predicate;
 pub use self::prover::Prover;
